backend/requirements.txt
Django>=5.0,<5.3
djangorestframework
djangorestframework-simplejwt
drf_spectacular
psycopg2-binary
django-filter
django-cors-headers
pytest-django
requests